Move web stuff to web package
