+++ mautrix_telegram/web/common/__init__.py
@@ -1,0 +1,1 @@
+from .auth_api import AuthAPI
