+++ mautrix_telegram/web/public/favicon.png
